split out FRSM-17
